Aprimoramento e Correções do projeto

+++ analise-credito-api/src/main/java/com/tqiprojeto/analisecreditoapi/service/ClienteService.java
@@ -1,9 +1,16 @@
 package com.tqiprojeto.analisecreditoapi.service;
 
 import com.tqiprojeto.analisecreditoapi.entity.Cliente;
+import com.tqiprojeto.analisecreditoapi.entity.Emprestimo;
 import com.tqiprojeto.analisecreditoapi.entity.Endereco;
+import com.tqiprojeto.analisecreditoapi.exception.ClienteDbException;
 import com.tqiprojeto.analisecreditoapi.exception.ClienteNaoCadastradoException;
+import com.tqiprojeto.analisecreditoapi.exception.EmprestimoNaoCadastradoException;
 import com.tqiprojeto.analisecreditoapi.repository.ClienteRepository;
+import com.tqiprojeto.analisecreditoapi.repository.EmprestimoRepository;
+import com.tqiprojeto.analisecreditoapi.repository.EmprestimoRetornoDetalhado;
+import com.tqiprojeto.analisecreditoapi.repository.EmprestimoRetornoSimples;
+import org.hibernate.JDBCException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,8 +18,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
+import javax.validation.ConstraintViolationException;
 import java.util.List;
-
 import java.util.Optional;
 
 
@@ -23,6 +30,7 @@ public class ClienteService {
     private ClienteRepository clienteRepository;
     private PasswordEncoder passwordEncoder;
 
+    private EmprestimoRepository emprestimoRepository;
 
     @Autowired
     public ClienteService(ClienteRepository clienteRepository, PasswordEncoder passwordEncoder) {
@@ -30,18 +38,21 @@ public class ClienteService {
         this.passwordEncoder = passwordEncoder;
     }
 
+    //método para listar todos os clientes
     public List<Cliente> listarTodos() {
 
         return clienteRepository.findAll();
 
     }
 
+    //método buscar um cliente por ID
     public Cliente buscarPorId(Integer id) throws ClienteNaoCadastradoException {
 
         Cliente cliente = clienteRepository.findById(id).orElseThrow(() -> new ClienteNaoCadastradoException(id));
         return cliente;
     }
 
+    //método para inserir um cliente
     public Cliente inserir(Cliente cliente) {
 
         setEnderecoCliente(cliente);
@@ -50,6 +61,7 @@ public class ClienteService {
         return clienteRepository.save(cliente);
     }
 
+    //método para atualizar um cliente cadastrado, informando o ID e demais atributos do cliente a ser atualizado
     public Cliente atualizar(Integer id, Cliente cliente) throws ClienteNaoCadastradoException {
 
         clienteRepository.findById(id).orElseThrow(() -> new ClienteNaoCadastradoException(id));
@@ -59,11 +71,22 @@ public class ClienteService {
         return clienteRepository.save(cliente);
     }
 
+    //método para excluir um cliente cadastrado
+    public void deletar(Integer id) throws ClienteNaoCadastradoException, ClienteDbException {
 
-    public void deletar(Integer id) throws ClienteNaoCadastradoException{
-        clienteRepository.deleteById(id);
+        verificarExistencia(id);
+
+        List<EmprestimoRetornoDetalhado> emprestimo = emprestimoRepository.listarEmprestimoDetalhado(id);
+
+            if (emprestimo.isEmpty()){
+                clienteRepository.deleteById(id);
+            }
+
+        else {throw new ClienteDbException(id);}
+
     }
 
+    //método para buscar um endereço automaticamente através do CEP (utilizado API do VIA CEP)
     private Endereco criarEndereco (String zipCode) {
 
         String url = "https://viacep.com.br/ws/"+zipCode+"/json/";
@@ -72,6 +95,7 @@ public class ClienteService {
         return endereco;
     }
 
+    //método para validar o login, verificado pelo email e senha do cliente cadastrado
     public ResponseEntity<Boolean> login (String email, String senha){
 
         Optional<Cliente> clienteEmail = clienteRepository.findByEmail(email);
@@ -88,6 +112,7 @@ public class ClienteService {
 
     }
 
+    //método usado para setar as informações do endereço do cliente (informações do endereço fornecidas pelo método criarEndereço)
     private Cliente setEnderecoCliente (Cliente cliente){
 
         Endereco endereco = criarEndereco(cliente.getCep());
@@ -95,10 +120,18 @@ public class ClienteService {
         return cliente;
     }
 
+    //método utilizado para criptografar a senha do cliente
     private Cliente encriptarSenha (Cliente cliente){
 
         cliente.setSenha(passwordEncoder.encode(cliente.getSenha()));
         return cliente;
     }
+
+    //método para everificar a existencia de cliente cadastrado para realizar exclusão
+    private Cliente verificarExistencia(Integer id) throws ClienteNaoCadastradoException {
+        return clienteRepository.findById(id)
+                .orElseThrow(() -> new ClienteNaoCadastradoException(id));
+    }
+
 
 }
